# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blotato Comments API",
    "description": "Retrieve comments for a published post and reply to a comment, across multiple social platforms.",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "accountContext": {
        "type": "apiKey",
        "name": "X-Account-Id",
        "in": "header",
        "description": "Tenant context supplied by the platform gateway that already authenticated the caller (assumption A-001)."
      }
    },
    "schemas": {
      "Comment": {
        "type": "object",
        "title": "Comment",
        "description": "A normalized comment. Provider identifiers are never exposed.",
        "required": [
          "id",
          "postId",
          "platform",
          "author",
          "body",
          "parentCommentId",
          "publishedAt",
          "updatedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "Service-owned identifier (ADR-0010)."
          },
          "postId": {
            "type": "string",
            "description": "Internal identifier of the published post."
          },
          "platform": {
            "type": "string",
            "enum": [
              "facebook",
              "instagram",
              "linkedin",
              "x",
              "youtube"
            ]
          },
          "author": {
            "type": "object",
            "required": [
              "id",
              "displayName"
            ],
            "additionalProperties": false,
            "properties": {
              "id": {
                "type": "string",
                "description": "The provider's author identifier."
              },
              "displayName": {
                "type": "string"
              },
              "profileUrl": {
                "type": "string"
              }
            }
          },
          "body": {
            "type": "string"
          },
          "parentCommentId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Parent comment, or null for a top-level comment."
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "title": "Pagination",
        "required": [
          "nextCursor",
          "hasMore"
        ],
        "additionalProperties": false,
        "properties": {
          "nextCursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "Opaque cursor for the next page. Clients must not decode or construct it."
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "Error": {
        "type": "object",
        "title": "Error",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "POST_NOT_FOUND",
                  "COMMENT_NOT_FOUND",
                  "IDEMPOTENCY_CONFLICT",
                  "UNSUPPORTED_CAPABILITY",
                  "PROVIDER_RATE_LIMITED",
                  "PROVIDER_ERROR",
                  "PROVIDER_UNAVAILABLE",
                  "INVALID_CURSOR",
                  "INVALID_REQUEST",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "description": "Correlates the response with the service log record."
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/v2/posts/{postId}/comments": {
      "get": {
        "operationId": "listComments",
        "summary": "Retrieve comments for a published post",
        "tags": [
          "comments"
        ],
        "description": "Answers from the local snapshot of the post, fetching a page from the provider when the snapshot cannot satisfy the requested position.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Number of comments requested."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Opaque cursor from a previous response."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "postId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A page of comments.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "A page of comments.",
                  "type": "object",
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be parsed or validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Caller credentials are missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "The resource is not visible in the caller scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The provider cannot perform the requested operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A provider or service rate limit was reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected failure occurred inside the service.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The provider returned an upstream failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The provider is temporarily unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v2/comments/{commentId}/replies": {
      "post": {
        "operationId": "replyToComment",
        "summary": "Publish a reply to a comment",
        "tags": [
          "comments"
        ],
        "description": "Publishes a reply through the post provider. The idempotency key makes a retry safe: a repeated request returns the reply already published rather than publishing a second one.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "body"
                ],
                "additionalProperties": false,
                "properties": {
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "commentId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "description": "Stable across retries of the same logical reply."
          }
        ],
        "responses": {
          "201": {
            "description": "The published reply.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The published reply.",
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Comment"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request cannot be parsed or validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Caller credentials are missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "The resource is not visible in the caller scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key cannot be honoured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The provider cannot perform the requested operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A provider or service rate limit was reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected failure occurred inside the service.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The provider returned an upstream failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The provider is temporarily unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "accountContext": []
    }
  ],
  "tags": [
    {
      "name": "comments",
      "description": "Comment retrieval and replies"
    }
  ]
}